app/tools/scc_tools.py
from google.cloud import securitycenter_v1

def list_active_findings(org_id: str, severity_filter: list[str], page_size: int = 1000):
    """Placeholder for listing findings - implemented in graph/ingestion/finding_ingester.py"""
    pass

def get_finding_detail(finding_id: str, org_id: str) -> dict:
    """Fetch details for a specific finding."""
    client = securitycenter_v1.SecurityCenterClient()
    # Stub implementation
    return {
        "finding_id": finding_id,
        "state": "ACTIVE",
        "severity": "HIGH",
        "resource_name": "//compute.googleapis.com/projects/test-project/zones/us-central1-a/instances/test-instance"
    }

def mute_resolved_finding(finding_id: str, org_id: str) -> None:
    """Mute a finding in SCC after successful remediation."""
    client = securitycenter_v1.SecurityCenterClient()
    # Logic to call SetMute API
    print(f"Muted finding {finding_id}")
